% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Title}
\usage{
elnet_coord(x, y, lambda, alpha, tol)
}
\arguments{
\item{x}{Predictor matrix x, with n rows and p columns}

\item{y}{Response vector y, with n elements}

\item{lambda}{Lasso tuning parameter}

\item{alpha}{Elastic net alpha parameter. elastic net simplifies to Ridge
regression under alpha = 0, and to Lasso regression under alpha = 1}

\item{tol}{Tolerance for convergence of betas. Algorithm will complete when
the maximum change in beta values is smaller than the tol value}
}
\value{
Matrix where each column is the p-length vector of beta estimates
for one iteration
}
\description{
Title
}
\examples{
beta = c(2,0,-2,0,1,0,-1,rep(0,13))
covar = diag(1,20)
covar[1,2] = covar[2,1] = covar[5,6] = covar[6,5] = .8
x = MASS::mvrnorm(20, rep(0, 20), covar)
y = x \%*\% beta + rnorm(20, 0, 1)

}
