% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{Soft thresholding for elnet_coord}
\usage{
soft(rx, lambda, alpha)
}
\arguments{
\item{rx}{rx object from elnet_coord function, product of partial residual
and a column of x}

\item{lambda}{lambda value specified in elnet_coord}

\item{alpha}{alpha value specified in elnet_coord}
}
\value{
Soft thresholded rx value
}
\description{
Soft thresholding for elnet_coord
}
