% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{Iterative solver for Linear Systems of Equations}
\usage{
solve_ols(A, b, method = "gs", iter = 5000, ncores = NULL)
}
\arguments{
\item{A}{Matrix with n rows and p columns}

\item{b}{Vector of length n}

\item{method}{Iterative method for solving Ax = b, either 'gs' for
Gauss-Seidel (sequential), 'jacobi' for Jacobi (sequential), or
'parallel' for parallel Jacobi.}

\item{iter}{Number of iterations for the solver.}

\item{ncores}{Optionally set number of cores. If left empty, solve_ols will
get the number of cores on the user's system and subtract 1.}
}
\value{
Numeric vector x, solution to Ax = b.
}
\description{
solve_ols can use Gauss-Seidel , Jacobi(sequential), or Jacobi
(parallel) to solve a system of equations Ax = b given A and b.
}
\examples{
A = diag(2, nrow = 20)
A[abs(row(A) -col(A)) == 1] = -1
b = A\%*\%(0.5*1:20)
solve_ols(A,b)
solve_ols(A,b, method = 'parallel', iter = 1000, ncores = 6)

}
